Moved winston logger to utils
