+++ src/index.js
@@ -1,5 +1,5 @@
 const app = require('./app')
-const logger = require('./config/winston')
+const logger = require('./utils/winston')
 
 const PORT = process.env.PORT || 3000
 const HOST = process.env.HOST || '0.0.0.0'
